Move single remaining class to power folder
