+++ src/main/java/crazypants/enderio/item/darksteel/ItemDarkSteelPickaxe.java
@@ -34,7 +34,7 @@ import crazypants.enderio.config.Config;
 import crazypants.enderio.item.darksteel.upgrade.EnergyUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.SpoonUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.TravelUpgrade;
-import crazypants.enderio.machine.power.PowerDisplayUtil;
+import crazypants.enderio.power.PowerDisplayUtil;
 import crazypants.enderio.teleport.TravelController;
 
 public class ItemDarkSteelPickaxe extends ItemPickaxe
