+++ src/main/java/crazypants/enderio/item/darksteel/upgrade/EnergyUpgrade.java
@@ -17,7 +17,7 @@ import crazypants.enderio.EnderIO;
 import crazypants.enderio.config.Config;
 import crazypants.enderio.item.darksteel.IDarkSteelItem;
 import crazypants.enderio.item.darksteel.IDarkSteelItem.IEndSteelItem;
-import crazypants.enderio.machine.power.PowerDisplayUtil;
+import crazypants.enderio.power.PowerDisplayUtil;
 import crazypants.enderio.material.Material;
 
 public class EnergyUpgrade extends AbstractUpgrade {
