+++ src/main/java/crazypants/enderio/machine/crafter/GuiCrafter.java
@@ -17,7 +17,7 @@ import crazypants.enderio.config.Config;
 import crazypants.enderio.gui.IconEIO;
 import crazypants.enderio.machine.PacketItemBuffer;
 import crazypants.enderio.machine.gui.GuiPoweredMachineBase;
-import crazypants.enderio.machine.power.PowerDisplayUtil;
+import crazypants.enderio.power.PowerDisplayUtil;
 import crazypants.enderio.network.PacketHandler;
 
 public class GuiCrafter extends GuiPoweredMachineBase<TileCrafter> {
