+++ src/main/java/crazypants/enderio/machine/generator/combustion/GuiCombustionGenerator.java
@@ -16,7 +16,7 @@ import crazypants.enderio.EnderIO;
 import crazypants.enderio.fluid.Fluids;
 import crazypants.enderio.machine.IoMode;
 import crazypants.enderio.machine.gui.GuiPoweredMachineBase;
-import crazypants.enderio.machine.power.PowerDisplayUtil;
+import crazypants.enderio.power.PowerDisplayUtil;
 
 public class GuiCombustionGenerator extends GuiPoweredMachineBase<TileCombustionGenerator> {
