+++ src/main/java/crazypants/enderio/machine/generator/stirling/GuiStirlingGenerator.java
@@ -19,7 +19,7 @@ import cpw.mods.fml.relauncher.SideOnly;
 import crazypants.enderio.EnderIO;
 import crazypants.enderio.config.Config;
 import crazypants.enderio.machine.gui.GuiPoweredMachineBase;
-import crazypants.enderio.machine.power.PowerDisplayUtil;
+import crazypants.enderio.power.PowerDisplayUtil;
 import crazypants.enderio.power.Capacitors;
 
 @SideOnly(Side.CLIENT)
