+++ src/main/java/crazypants/enderio/nei/AlloySmelterRecipeHandler.java
@@ -21,7 +21,7 @@ import codechicken.nei.recipe.TemplateRecipeHandler;
 import crazypants.enderio.gui.IconEIO;
 import crazypants.enderio.machine.alloy.AlloyRecipeManager;
 import crazypants.enderio.machine.alloy.GuiAlloySmelter;
-import crazypants.enderio.machine.power.PowerDisplayUtil;
+import crazypants.enderio.power.PowerDisplayUtil;
 import crazypants.enderio.machine.recipe.IRecipe;
 import crazypants.enderio.machine.recipe.RecipeInput;
